requirements.txt
streamlit
streamlit-authenticator>=0.3.1
pyyaml